Refactor data/modules into project-modules and skill-modules
